Azure Sql DB,Key Vault,Storage account implementation

+++ Assignment_UKHO.Data/Migrations/AppDbContextModelSnapshot.cs
@@ -150,8 +150,8 @@ namespace Assignment_UKHO.Data.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("FileSize")
-                        .HasColumnType("int");
+                    b.Property<long>("FileSize")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("Hash")
                         .IsRequired()
